requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langdetect==1.0.9
MarkupSafe==2.1.5
multidict==6.1.0
mutagen==1.47.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
twilio==9.3.0
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
